% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpack.R
\name{lift_up}
\alias{lift_up}
\title{Change function to accept a packed vector as argument}
\usage{
lift_up(f, up)
}
\arguments{
\item{f}{a function}

\item{up}{an unpacking method}
}
\value{
a function
}
\description{
Similar to the family of lift functions in purrr. An example works better than an explanation, so please see below.
}
\examples{
f <- function(a,b) a + b #A function of two arguments
l <- list(a= matrix(1:4,2,2),b = 5)
up <- gen.vunpack(l)
f.lifted <- lift_up(f,up)
f(l$a,l$b)
f.lifted(vpack(l))
}
\author{
Simon Barthelme
}

