% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpack.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator from magrittr package}
\arguments{
\item{lhs}{data}

\item{rhs}{a function}
}
\description{
The pipe operator lets you string function calls together. See help from magrittr package for more.
}

