% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpack.R
\name{vpack}
\alias{vpack}
\title{Pack a list of objects into a vector}
\usage{
vpack(l)
}
\arguments{
\item{l}{a list of objects or a single object}
}
\value{
a vector
}
\description{
Turns a list of objects into a concatenated vector. Objects must have an "as.vector" method.
If given a single object, vpack just runs as.vector.
}
\examples{
list(a=1:3,b=matrix(1:4,2,2)) \%>\% vpack
matrix(1:4,2,2) \%>\% vpack
}
\author{
Simon Barthelme
}
\seealso{
gen.vunpack
}

