% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpack.R
\name{vpgrad}
\alias{vpgrad}
\title{An interface to numDeriv::grad for numerical gradients over multiple arguments}
\usage{
vpgrad(fun, x, ...)
}
\arguments{
\item{fun}{function to differentiate (the names of the arguments should be the *same ones* as in the "guess" argument)}

\item{x}{compute gradient at x}

\item{...}{passed on to grad}
}
\description{
Same as vpoptim, pretty much. See examples
}
\examples{
#gradient of a function of two arguments
vpgrad(function(x,y) x+y,list(x=1,y=3))
#gradient of a matrix function, returned as a matrix
vpgrad(det,diag(2))
}

